logscope: use long for _level, because level is unsigned int most other places
SILENT level does not get properly restored because of overflow issues

+++ streamlog/source/include/streamlog/logscope.h
@@ -65,7 +65,7 @@ namespace streamlog{
   protected: 
     logstream* _ls ;
     std::string _name ;
-    int _level;
+    long _level;
     logscope() {}
     
   };
